ml/requirements.txt
pandas
numpy
scikit-learn
tensorflow-macos
tensorflow-metal
pyarrow
